Compact Story Loading Into One File
Moves loading resolution into a per-story yaml file instead of hardcoding in the script. This can eventually be abstracted into assembling the subcomponents from multiple files
Off-repo, done a lot of work on simplifying and unifying the story prompts. Making work on identifying the relevant concepts for generalizing the architect