New Specfinding Approach Using Prewrite Essay
Identified a new approach for specfinding by writing an entire essay of the plot idea ahead of time and then extracting the spec document from that. This seems to work really well in practice, but still doing some work with finetuning. There will still need to be a followup conversation that makes sure the auto-detected spec is okay, but that will only require a slight modification of the orchestrator